+++ src/libaktualizr/opcuabridge/opcuabridge_secondary.cc
@@ -110,5 +110,5 @@ int main(int argc, char* argv[]) {
   UA_Server_delete(server);
   UA_ServerConfig_delete(config);
 
-  return retval;
+  return static_cast<int>(retval);
 }
